backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
opencv-python==4.10.0.84
torch==2.4.1
torchvision==0.19.1
mediapipe==0.10.14
numpy==2.1.1
scipy==1.14.1
Pillow==10.4.0
transformers==4.44.2
librosa==0.10.2
openai-whisper==20231117